requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
